Refactor Form away from AppNavigator
