+++ Mindset/Mobile/SupportingFiles/AppDelegate.swift
@@ -14,15 +14,6 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
         window = UIWindow(frame: UIScreen.main.bounds)
         window?.rootViewController = navigationController
         window?.makeKeyAndVisible()
-
-        let realm = try! Realm()
-
-        let realmRecords = realm.objects(RealmRecordEntry.self)
-        print("**********")
-        print(realmRecords)
-        print("**********")
-        
-
         return true
     }
